Cast si_drv2 to intptr_t instead of int to quiet warnings.
Tested on:	alpha
Reported by:	marcel (on an ia64)


Former-commit-id: 4e56dd76f8205ca5fa6d9ac9ea5c2689a54e6a0d

+++ sys/dev/rc/rc.c
@@ -242,7 +242,7 @@ name|CALLOUT
 parameter_list|(
 name|dev
 parameter_list|)
-value|(((int)(dev)->si_drv2) != 0)
+value|(((intptr_t)(dev)->si_drv2) != 0)
 end_define
 
 begin_comment
